fix: services cleanup
